course names and courseId

+++ backend/Services/StudentServices/StudentService.cs
@@ -1,20 +1,23 @@
 ﻿using AutoMapper;
 using backend.DTOs.StudentDtos;
 using backend.Models;
+using backend.Repositories.StudentEnrollmentRepo;
 using backend.Repositories.StudentRepo;
+using Microsoft.IdentityModel.Tokens;
 
 namespace backend.Services.StudentServices
 {
 
     public class StudentService : IStudentService
-    {
+    {   private readonly IEnrollmentRepository _enrollmentRepository;
         private readonly IStudentRepository _studentRepository;
         private readonly IMapper _mapper;
 
-        public StudentService(IStudentRepository studentRepository, IMapper mapper)
+        public StudentService(IStudentRepository studentRepository, IMapper mapper , IEnrollmentRepository enrollmentRepository)
         {
             _studentRepository = studentRepository;
             _mapper = mapper;
+            _enrollmentRepository = enrollmentRepository;
         }
 
         public async Task<StudentResponseDto> CreateStudentAsync(CreateStudentDto createDto)
@@ -26,6 +29,11 @@ namespace backend.Services.StudentServices
 
         public async Task<bool> DeleteStudentAsync(string studentId)
         {
+            var data = await _enrollmentRepository.GetCourseIdsByStudentIdAsync(studentId);
+            if(data.Count > 0)
+            {
+                throw new Exception("Student is already assigned to courses!");
+            }
             return await _studentRepository.DeleteAsync(studentId);
         }
